+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.1 - 08/02/21
 - Standaridation of using 'name' attribute over 'message' attribute for multiple classes
 - Basic implementation of ReplicationAgent class (w/ .cpp and .h files)
+- Modifications of other files to account for use of this new class
 
 ## 0.2.0 - 06/02/21
 - Minor console output cleanup